services/orchestrator/requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
httpx
slack_sdk
PyGithub